+++ NEWS.md
@@ -10,6 +10,7 @@ The latest release of the SAMtool package is available on [CRAN](https://CRAN.R-
 - RCM can fit to list of biological inputs `StockPars` instead of an operating model with `RCM(StockPars, RCMdata)`. See `StockPars` section in `help("RCM")`.
 - Add partial bias correction to the initial abundance at age vector in `RCM` with argument `pbc_early_recdev`. The bias correction from logspace to normal space is `exp(log_early_rec_dev[a] - 0.5 * pbc_recdev[a] * sigmaR^2)` if the year-specific rec dev is estimated.
 - `RCMdata@I_delta` (survey timing within time step) can now be a matrix by year and survey.
+- Fix typo in `Shortcut` function
 
 ## SAMtool 1.7.0
 - Add argument to adjust annual recruitment deviation bias correction in `RCM` in argument `pbc_recdev`. The bias correction from logspace to normal space is `exp(log_rec_dev[y] - 0.5 * pbc_recdev[y] * sigmaR^2)` if the year-specific rec dev is estimated.